requirements.txt
django~=5.1.2
psycopg2-binary
djangorestframework~=3.15.2
djangorestframework-simplejwt
celery
redis